// src/types/recipe.d.ts
export interface Cocktail {
  idDrink: string;
  strDrink: string;
  strDrinkThumb: string;
  strInstructions?: string;
  strCategory?: string;
  // Add other properties you might use
}

export interface Meal {
  idMeal: string;
  strMeal: string;
  strMealThumb: string;
  strInstructions?: string;
  strCategory?: string;
  // Add other properties you might use
}